+++ src/Vector3.java
@@ -1,5 +1,5 @@
 public class Vector3{
-    double x,y,z;
+    float x,y,z;
 
 
     public Vector3(){
@@ -10,7 +10,7 @@ public class Vector3{
 
     }
 
-    public Vector3(double x,double y,double z){
+    public Vector3(float x,float y,float z){
 
         this.x = x;
         this.y = y;
@@ -28,9 +28,9 @@ public class Vector3{
     }
 
     public void normalize() {
-        double length;
+        float length;
 
-        length = Math.sqrt(x*x + y*y + z*z);
+        length = (float) Math.sqrt(x*x + y*y + z*z);
 
         try {
             x = x / length;
@@ -42,19 +42,19 @@ public class Vector3{
         }
     }
 
-    public double dotProduct(Vector3 v) {
+    public float dotProduct(Vector3 v) {
         return((x*v.x + y*v.y + z*v.z));
     }
 
 
     public void crossProduct(Vector3 v) {
-        double tmpx = y*v.z - z*v.y,
+        float tmpx = y*v.z - z*v.y,
                 tmpy = z*v.x - x*v.z,
                 tmpz = x*v.y - y*v.x;
         x = tmpx; y = tmpy; z = tmpz;
     }
 
-    public void mult(double factor) {
+    public void mult(float factor) {
         x=x*factor; y=y*factor; z=z*factor;
     }
 
@@ -74,8 +74,8 @@ public class Vector3{
 
 
     }
-    public double distance (Vector3 v){
-        double d = Math.sqrt(Math.pow(v.x- this.x,2) + Math.pow(v.y- this.y,2) + Math.pow(v.z- this.z,2));
+    public float distance (Vector3 v){
+        float d = (float)(Math.sqrt(Math.pow(v.x- this.x,2) + Math.pow(v.y- this.y,2) + Math.pow(v.z- this.z,2)));
         return d;
 
     }
